pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "any-router"
version = "0.1.0"
description = "极简个人意图路由与自动化 Agent"
readme = "项目文档.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Any-Router" },
]
dependencies = [
    "requests>=2.28",
    "python-dotenv>=1.0",
]

[project.scripts]
any-router = "any_router.cli:main"

[tool.setuptools.packages.find]
include = ["any_router*"]
